requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
anyio==4.9.0
asyncclick==8.1.8.0
attrs==25.3.0
certifi==2025.7.14
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
docopt==0.6.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
mashumaro==3.16
multidict==6.6.3
pipreqs==0.4.13
propcache==0.3.2
pycparser==2.22
python-dotenv==1.1.1
python-kasa==0.10.2
python-telegram-bot==22.3
requests==2.32.4
scapy==2.6.1
sniffio==1.3.1
typing_extensions==4.14.1
urllib3==2.5.0
yarg==0.1.10
yarl==1.20.1
